+++ test/lib/absinthe/relay/node/parse_ids_test.exs
@@ -17,7 +17,7 @@ defmodule Absinthe.Relay.Node.ParseIDsTest do
 
   defmodule Schema do
     use Absinthe.Schema
-    use Absinthe.Relay.Schema
+    use Absinthe.Relay.Schema, :classic
 
     alias Absinthe.Relay.Node.ParseIDsTest.Foo
     alias Absinthe.Relay.Node.ParseIDsTest.Parent
@@ -66,7 +66,7 @@ defmodule Absinthe.Relay.Node.ParseIDsTest do
     end
 
     input_object :child_input do
-      field :id, non_null(:id)
+      field :id, :id
     end
 
     query do
@@ -128,6 +128,9 @@ defmodule Absinthe.Relay.Node.ParseIDsTest do
     defp resolve_foo(%{foo_id: id}, _) do
       {:ok, Map.get(@foos, id)}
     end
+    defp resolve_foo(%{foobar_id: nil}, _) do
+      {:ok, nil}
+    end
     defp resolve_foo(%{foobar_id: %{id: id, type: :foo}}, _) do
       {:ok, Map.get(@foos, id)}
     end
@@ -138,7 +141,30 @@ defmodule Absinthe.Relay.Node.ParseIDsTest do
     end
 
     defp resolve_parent(args, _) do
-      {:ok, args}
+      {:ok, args |> to_parent_output}
+    end
+
+    # This is just a utility that converts the input value into the
+    # expected output value (which has non-null constraints).
+    #
+    # It doesn't have any value outside these tests!
+    #
+    defp to_parent_output(%{id: nil}) do
+      nil
+    end
+    defp to_parent_output(values) when is_list(values) do
+      for value <- values do
+        value
+        |> to_parent_output
+      end
+    end
+    defp to_parent_output(%{} = args) do
+      for {key, value} <- args, into: %{} do
+        {key, value |> to_parent_output}
+      end
+    end
+    defp to_parent_output(value) do
+      value
     end
 
     @update_parent_ids {
@@ -166,90 +192,179 @@ defmodule Absinthe.Relay.Node.ParseIDsTest do
   @foo1_id Base.encode64("Foo:1")
   @foo2_id Base.encode64("Foo:2")
 
-  it "parses one id correctly" do
-    result =
-      """
-      {
-        foo(fooId: "#{@foo1_id}") {
-          id
-          name
-        }
-      }
-      """
-      |> Absinthe.run(Schema)
-    assert {:ok, %{data: %{"foo" => %{"name" => "Foo 1", "id" => @foo1_id}}}} == result
-  end
+  describe "parses one id" do
 
-  it "parses a list of ids correctly" do
-    result =
-      """
-      {
-        foos(fooIds: ["#{@foo1_id}", "#{@foo2_id}"]) { id name }
-      }
-      """
-      |> Absinthe.run(Schema)
-    assert {:ok,
-      %{
-        data: %{
-          "foos" => [
-            %{"name" => "Foo 1", "id" => @foo1_id},
-            %{"name" => "Foo 2", "id" => @foo2_id}
-          ]
-        }
-      }
-    } == result
-  end
-
-  it "parses an id into one of multiple node types" do
-    result =
-      """
-      {
-        foo(foobarId: "#{@foo1_id}") { id name }
-      }
-      """
-      |> Absinthe.run(Schema)
-    assert {:ok, %{data: %{"foo" => %{"name" => "Foo 1", "id" => @foo1_id}}}} == result
-  end
-
-  @tag :focus
-  it "parses nested ids" do
-    encoded_parent_id = Base.encode64("Parent:1")
-    encoded_child1_id = Base.encode64("Child:1")
-    encoded_child2_id = Base.encode64("Child:1")
-    result =
-      """
-      mutation Foobar {
-        updateParent(input: {
-          clientMutationId: "abc",
-          parent: {
-            id: "#{encoded_parent_id}",
-            children: [{ id: "#{encoded_child1_id}"}, {id: "#{encoded_child2_id}"}],
-            child: { id: "#{encoded_child2_id}"}
-          }
-        }) {
-          parent {
+    test "succeeds with a non-null value" do
+      result =
+        """
+        {
+          foo(fooId: "#{@foo1_id}") {
             id
-            children { id }
-            child { id }
-            }
+            name
           }
-      }
-      """
-      |> Absinthe.run(Schema)
-
-    expected_parent_data = %{
-      "parent" => %{
-        "id" => encoded_parent_id, # The output re-converts everything to global_ids.
-        "children" => [%{"id" => encoded_child1_id}, %{"id" => encoded_child2_id}],
-        "child" => %{
-          "id" => encoded_child2_id
         }
-      }
-    }
-    assert {:ok, %{data: %{"updateParent" => expected_parent_data}}} == result
+        """
+        |> Absinthe.run(Schema)
+      assert {:ok, %{data: %{"foo" => %{"name" => "Foo 1", "id" => @foo1_id}}}} == result
+    end
+
+    test "succeeds with a null value" do
+      result =
+        """
+        {
+          foo(fooId: null) {
+            id
+            name
+          }
+        }
+        """
+        |> Absinthe.run(Schema)
+      assert {:ok, %{data: %{"foo" => nil}}} == result
+    end
+
   end
 
-  it "parses incorrect nested ids" do
+  describe "parses a list of ids" do
+
+    test "succeeds with a non-null value" do
+      result =
+        """
+        {
+          foos(fooIds: ["#{@foo1_id}", "#{@foo2_id}"]) { id name }
+        }
+        """
+        |> Absinthe.run(Schema)
+      assert {:ok,
+        %{
+          data: %{
+            "foos" => [
+              %{"name" => "Foo 1", "id" => @foo1_id},
+              %{"name" => "Foo 2", "id" => @foo2_id}
+            ]
+          }
+        }
+      } == result
+    end
+
+    test "succeeds with a null value" do
+      result =
+        """
+        {
+          foos(fooIds: [null, "#{@foo2_id}"]) { id name }
+        }
+        """
+        |> Absinthe.run(Schema)
+      assert {:ok,
+              %{
+                data: %{
+                  "foos" => [
+                    nil,
+                    %{"name" => "Foo 2", "id" => @foo2_id}
+                  ]
+                }
+              }
+      } == result
+    end
+
+  end
+
+  describe "parsing an id into one of multiple node types" do
+    test "parses an non-null id into one of multiple node types" do
+      result =
+        """
+        {
+          foo(foobarId: "#{@foo1_id}") { id name }
+        }
+        """
+        |> Absinthe.run(Schema)
+      assert {:ok, %{data: %{"foo" => %{"name" => "Foo 1", "id" => @foo1_id}}}} == result
+    end
+    test "parses null" do
+      result =
+        """
+        {
+          foo(foobarId: null) { id name }
+        }
+        """
+        |> Absinthe.run(Schema)
+      assert {:ok, %{data: %{"foo" => nil}}} == result
+    end
+  end
+
+  describe "parsing nested ids" do
+    test "works with non-null values" do
+      encoded_parent_id = Base.encode64("Parent:1")
+      encoded_child1_id = Base.encode64("Child:1")
+      encoded_child2_id = Base.encode64("Child:1")
+      result =
+        """
+        mutation Foobar {
+          updateParent(input: {
+            clientMutationId: "abc",
+            parent: {
+              id: "#{encoded_parent_id}",
+              children: [{ id: "#{encoded_child1_id}"}, {id: "#{encoded_child2_id}"}],
+              child: { id: "#{encoded_child2_id}"}
+            }
+          }) {
+            parent {
+              id
+              children { id }
+              child { id }
+              }
+            }
+        }
+        """
+        |> Absinthe.run(Schema)
+
+      expected_parent_data = %{
+        "parent" => %{
+          "id" => encoded_parent_id, # The output re-converts everything to global_ids.
+          "children" => [%{"id" => encoded_child1_id}, %{"id" => encoded_child2_id}],
+          "child" => %{
+            "id" => encoded_child2_id
+          }
+        }
+      }
+      assert {:ok, %{data: %{"updateParent" => expected_parent_data}}} == result
+    end
+    test "works with null leaf values" do
+      encoded_parent_id = Base.encode64("Parent:1")
+      encoded_child1_id = Base.encode64("Child:1")
+      result =
+        """
+        mutation Foobar {
+          updateParent(input: {
+            clientMutationId: "abc",
+            parent: {
+              id: "#{encoded_parent_id}",
+              children: [{ id: "#{encoded_child1_id}" }, { id: null }],
+              child: { id: null }
+            }
+          }) {
+            parent {
+              id
+              children { id }
+              child { id }
+              }
+            }
+        }
+        """
+        |> Absinthe.run(Schema)
+
+      expected_parent_data = %{
+        "parent" => %{
+          "id" => encoded_parent_id, # The output re-converts everything to global_ids.
+          "children" => [%{"id" => encoded_child1_id}, nil],
+          "child" => nil
+        }
+      }
+      assert {:ok, %{data: %{"updateParent" => expected_parent_data}}} == result
+    end
+
+  end
+
+  test "parses incorrect nested ids" do
     encoded_parent_id = Base.encode64("Parent:1")
     incorrect_id = Node.to_global_id(:other_foo, 1, Schema)
     mutation =
@@ -274,12 +389,12 @@ defmodule Absinthe.Relay.Node.ParseIDsTest do
       data: %{"updateParent" => nil},
       errors: [%{
         locations: [%{column: 0, line: 2}],
-        message: ~s<In field "updateParent": In argument "input": In field "parent": In field "child": In field "id": Expected node type in ["Child"], found "FancyFoo".>
+        message: ~s<In argument "input": In field "parent": In field "child": In field "id": Expected node type in ["Child"], found "FancyFoo".>
       }]
     } = result
   end
 
-  it "handles one incorrect id correctly" do
+  test "handles one incorrect id correctly" do
     result =
       """
       {
@@ -294,13 +409,13 @@ defmodule Absinthe.Relay.Node.ParseIDsTest do
       :ok, %{
         data: %{},
         errors: [
-          %{message: ~s<In field "foo": In argument "fooId": Expected node type in ["Foo"], found "FancyFoo".>}
+          %{message: ~s<In argument "fooId": Expected node type in ["Foo"], found "FancyFoo".>}
         ]
       }
     } = result
   end
 
- it "parses nested ids with local middleware" do
+ test "parses nested ids with local middleware" do
     encoded_parent_id = Base.encode64("Parent:1")
     encoded_child1_id = Base.encode64("Child:1")
     encoded_child2_id = Base.encode64("Child:1")
@@ -311,7 +426,7 @@ defmodule Absinthe.Relay.Node.ParseIDsTest do
           clientMutationId: "abc",
           parent: {
             id: "#{encoded_parent_id}",
-            children: [{ id: "#{encoded_child1_id}"}, {id: "#{encoded_child2_id}"}],
+            children: [{ id: "#{encoded_child1_id}"}, {id: "#{encoded_child2_id}"}, {id: null}],
             child: { id: "#{encoded_child2_id}"}
           }
         }) {
@@ -328,7 +443,7 @@ defmodule Absinthe.Relay.Node.ParseIDsTest do
     expected_parent_data = %{
       "parent" => %{
         "id" => encoded_parent_id, # The output re-converts everything to global_ids.
-        "children" => [%{"id" => encoded_child1_id}, %{"id" => encoded_child2_id}],
+        "children" => [%{"id" => encoded_child1_id}, %{"id" => encoded_child2_id}, nil],
         "child" => %{
           "id" => encoded_child2_id
         }