+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+# 1.4.0
+
+Status: RC
+
+- Feature: Support `null` values in `ParseIDs` middleware (passed through as `nil` args)
+- Bug Fix: Support `null` values for `before` and `after` pagination arguments (expected by Relay Modern)
+
 ## 1.3.6
 - Type Spec Fix: Relax type constraints around `Connection.from_query`